+++ recipes/compute_base_config.rb
@@ -113,3 +113,6 @@ shared_dir_array.each do |dir|
     retry_delay 5
   end
 end
+
+# IMDS
+include_recipe 'aws-parallelcluster::imds_config'